requirements.txt
ansicon==1.89.0
arrow==1.2.1
asgiref==3.4.1
blessed==1.19.0
Django==4.0
django-picklefield==3.0.1
django-q==1.3.9
jinxed==1.1.0
python-dateutil==2.8.2
python-decouple==3.5
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
wcwidth==0.2.5
